Image data any

+++ plugins/input/gdal/gdal_datasource.cpp
@@ -109,7 +109,7 @@ gdal_datasource::gdal_datasource(parameters const& params)
     nbands_ = dataset->GetRasterCount();
     width_ = dataset->GetRasterXSize();
     height_ = dataset->GetRasterYSize();
-    desc_.add_descriptor(mapnik::attribute_descriptor("nodata", mapnik::Integer));
+    desc_.add_descriptor(mapnik::attribute_descriptor("nodata", mapnik::Double));
 
     double tr[6];
     bool bbox_override = false;
